// commandertracker.client/src/app/core/enums/colorIdentity.enum.ts
export enum ColorIdentity {
    Colorless,
    MonoWhite,
    MonoBlue,
    MonoBlack,
    MonoRed,
    MonoGreen,
    Azorius,
    Dimir,
    Rakdos,
    Gruul,
    Selesnya,
    Orzhov,
    Izzet,
    Golgari,
    Boros,
    Simic,
    Esper,
    Grixis,
    Jund,
    Naya,
    Bant,
    Abzan,
    Jeskai,
    Sultai,
    Mardu,
    Temur,
    YoreTiller,
    GlintEye,
    DuneBrood,
    InkTreader,
    WitchMaw,
    FiveColor
}

export enum ColorIdentityString {
    Colorless = '',
    MonoWhite = 'w',
    MonoBlue = 'u',
    MonoBlack = 'b',
    MonoRed = 'r',
    MonoGreen = 'g',
    Azorius = 'wu',
    Dimir = 'ub',
    Rakdos = 'br',
    Gruul = 'rg',
    Selesnya = 'gw',
    Orzhov = 'wb',
    Izzet = 'ur',
    Golgari = 'bg',
    Boros = 'rw',
    Simic = 'gu',
    Esper = 'wub',
    Grixis = 'urb',
    Jund = 'rbg',
    Naya = 'rgw',
    Bant = 'gwu',
    Abzan = 'wbg',
    Jeskai = 'urw',
    Sultai = 'bgu',
    Mardu = 'rwb',
    Temur = 'gur',
    YoreTiller = 'wubr',
    GlintEye = 'ubrg',
    DuneBrood = 'brgw',
    InkTreader = 'rgwu',
    WitchMaw = 'gwub',
    FiveColor = 'wubrg'
}

export enum ColorIdentityStringWubrgOrder {
    Colorless = '',
    MonoWhite = 'w',
    MonoBlue = 'u',
    MonoBlack = 'b',
    MonoRed = 'r',
    MonoGreen = 'g',
    Azorius = 'wu',
    Dimir = 'ub',
    Rakdos = 'br',
    Gruul = 'rg',
    Selesnya = 'wg',
    Orzhov = 'wb',
    Izzet = 'ur',
    Golgari = 'bg',
    Boros = 'wr',
    Simic = 'ug',
    Esper = 'wub',
    Grixis = 'ubr',
    Jund = 'brg',
    Naya = 'wrg',
    Bant = 'wug',
    Abzan = 'wbg',
    Jeskai = 'wur',
    Sultai = 'ubg',
    Mardu = 'wbr',
    Temur = 'urg',
    YoreTiller = 'wubr',
    GlintEye = 'ubrg',
    DuneBrood = 'wbrg',
    InkTreader = 'wurg',
    WitchMaw = 'wubg',
    FiveColor = 'wubrg'
}